Enter stage 1.5 bootloader
Similar to GRUB, store more of the bootloader code in the 62 sectors
immediately following the MBR (since partitions have to be
cylinder-aligned, this space is fine to use on a disk). Use the 0x13
BIOS function to copy these into memory immediately following the
bootloader.

Also add some hex printing functions for integers and a function to
dump registers.