Detect and print memory map
Query BIOS INT 0x15, AX=0xE820 to get the extended memory map. This
will be useful not only for the kernel but so that we can dynamically
allocate the page tables (which will be ~1MB for a 1GB HHDM). Also
added some notes on paging and simple protected mode printing
functions (which can't use the BIOS INT 0x10 function anymore).