Allocate memory for page table
This uses a (very dumb) allocator since there is only a single dynamic
allocation going on in the kernel (~1MB). Also add bootloader entries
to the memory map (which overlap the usable memory region they're in;
it's up to the kernel to normalize the memory map items).

Also some misc. cleanup, e.g. use the `.set` directive for constants
in assembly code.