Enable paging (1MB lower direct map and 1GB HHDM)
This sets up and enables two virtual memory regions via paging:

| Virtual base | Physical base | Length     | Notes                  |
| ------------ | ------------- | ---------- | ---------------------- |
| 0x1000       | 0x1000        | 0xFF000    | Low-memory direct map  |
| 0xC0000000   | 0x0           | 0x40000000 | Higher-half direct map |

The former is only required so that we don't have to change bootloader
memory addresses, and can be discarded once we begin executing in the
higher-half direct map. The latter is the main kernel memory virtual
address space.

The former is allocated using regular 4KB mappings (since we don't
want to map the null page), while the latter is more easily done via
4MB hugepages (with PSE). Note that we require PSE and PAE capability
(although the bootloader currently doesn't use the latter, it probably
makes sense in the future and it's similar to 64-bit paging).

At this point the bootloader has done pretty much all the setup I've
planned to do. The only thing left is to load the kernel -- the
details are still in the air.