install_bootloader.py to install bootloader/kernel onto disk image
This script copies the bootloader and kernel onto the same disk
image. The kernel will be copied to the first MBR partition so the
bootloader can easily find it.

The logic to actually parse this partition info and load the ELF file
isn't done yet, nor have we compiled a kernel image yet. This will
follow soon.