Kernel test framework
The approach I took in laos was not so good, so I'd like to
reformulate it here. There, you would compile the kernel with
`-DRUNTEST=<test_selection_criteria`, and this would only compile in
the tests you want. This means that running a different set of tests
means recompiling the kernel.

There's no reason this can't be done at runtime, so the new idea is to
build a test runner kernel that reads the test selection at runtime
(via serial port). The test framework has simple support for C++
namespaces.

Also implement a simple `string_view` implementation for the test
selection matcher, and write some tests for existing components.