Move kernel stack to a separate page
This solves a (huge) performance regression for all bootloader code
and anything running on the initial kernel thread. (Later kernel
threads are not affected as their stacks are allocated on their own
pages.) In particular, this fixed the issue that caused slow page
frame table initialization. See notes/hmm.md for the investigation
that led to this change.

Also moved the kernel data section to be page-aligned for good
measure, but this did not have any noticeable impact.