Makefile rules for debugging with symbols
The bootloader doesn't understand ELF file format so we normally
compile a flat binary file. This change adds Makefile rules to also
build an ELF file with debug symbols for GDB. The `runi` ("run
interactive") target runs QEMU with gdbserver, and the `gdb` target
connects to gdbserver and loads the debug symbols.