Write kernel to FAT32 partition + bootloader FAT32 support
The flat kernel binary file is now stored in a FAT32 partition, and
minimal FAT32 support is added to the bootloader to support finding
and loading the kernel. This is in preparation for loading user-space
binaries from the kernel, which shouldn't be stored in the kernel
binary itself -- FAT32 seemed like a simple fs choice for now.

We still have a ways to go before loading a user-space program from a
file -- we'll need at least a (SATA?) disk driver and a FAT32 fs
driver in the kernel. Eventually we'll probably also want a bona fide
ELF loader (also in the bootloader) so we can set permissions
correctly for different memory sections, and a virtual filesystem
layer (VFS) -- fun!