Kernel FAT32 driver
This replicates the basic implementation from the bootloader, and adds
support for printing 8.3-format filenames and subdirectories. This
exposes an interface for reading files and directories modelled after
the `read` and `getdents` syscalls (though these are intended to be at
a lower-level than filesystem syscalls, which would be exposed by the
VFS). I plan to keep the bootloader implementation minimal, hence the
separate implementation in the kernel.