GDT reinitialization in the kernel
I previously made the silly decision to reuse the GDT initialized by
the bootloader in the kernel. This (1) requires some workarounds to
live in non-bootloader-reclaimable memory and (2) doesn't make much
sense as the requirements differ. The bootloader needs ring0 segments
and an unreal mode descriptor. The kernel needs ring0, ring3, and TSS
segments.

As a result, add GDT initialization logic in the kernel. This
unfortunately duplicates more code from the bootloader but it keeps
the bootloader simple.