Virtual filesystem (VFS) layer
This expands the current simplistic FAT32 driver so that operations
happen via the File interface (similar to Linux's `struct
file`). Under the hood, this keeps a reference to a Dentry (`struct
dentry`), which has an underlying Inode (`struct inode`). Dentries
live in a hashtable cache for quick repeated retrieval.

This sets up the foundation for the process abstraction, since many
syscalls directly translate to VFS operations.

This is not thread-safe yet.