Implement process abstraction and jump into userspace!
This introduces the process abstraction that is core to timesharing.
Processes contain the following resources:
- A single schedulable thread
- A kernel stack
- A userspace stack
- Its own virtual address space (page table)
- Data/BSS/text region mappings from an ELF file

Process creation logic entails the following:
- Compiling userspace binaries (new Makefile)
- Creating new page table with cloned kernel mappings
- ELF loader
- mmap/munmap logic, including new VMA struct
- Jump to userspace

When switching to a thread associated with a userspace thread, the
scheduler will also switch the virtual address space. Two new
process-aware ISRs were implemented: a page-fault handler, and a
syscall handler (interrupt 0x80). Currently, the only syscall
implemented is _exit, which is also used to kill procs on segfault (in
lieu of the signal(7) mechanism).

Still not thread-safe